reinitialize logger for uploads
