+++ changelog/unreleased/extract-log-initialization.md
@@ -1,0 +1,5 @@
+Enhancement: extract log initialization
+
+To prepare reinitializing a logger for uploads we refactored the loginitialization into its own package
+
+https://github.com/cs3org/reva/pull/4368
